Move `hanoiTypes.ts` into `types/` directory
