+++ src/types/hanoiTypes.ts
