+++ src/utilities/getTowerContent.unit.test.ts
@@ -1,4 +1,4 @@
-import { HanoiTower } from "../components/Game/types";
+import { HanoiTower } from "../types";
 import { getTowerContent } from "./getTowerContent";
 
 describe("getTowerContent", () => {
